refactor(core): split proficiencies and make grants.resolve a leaf
Extract proficiency_collect and subclass apply; move feat token merge
into feats.grant_merge so grants.resolve no longer imports feats.
