+++ ui/menus/creation/subclass_picks.py
@@ -2,11 +2,13 @@
 
 from core.character.models import Character
 from core.mechanics.proficiencies import (
-    apply_subclass_proficiencies_to_character,
     is_valid_tool_selection,
     merge_proficiency_tokens,
 )
 from core.progression.class_progression import subclass_skill_picks_pending
+from core.progression.subclass_proficiencies import (
+    apply_subclass_proficiencies_to_character,
+)
 from core.types import LanguageCode, StringsDict
 from ui.menus.creation.expertise import apply_pending_expertise
 from ui.menus.creation.proficiencies import _pick_tools
